health: move formatConditions to HealthExamination and harden value parsing
Addresses review feedback on #16609:

- Move formatConditions from MyHealth's companion to HealthExamination's
  existing companion, which already owns the conditions field and reads/
  writes the blob in fromJson/serialize. Puts the read next to the write
  and drops the MyHealth import the adapter gained. MyHealth.kt is back to
  its pre-PR holder-only form.
- Switch the predicate from conditionsMap[it].asBoolean to
  JsonUtils.getBoolean(it, conditionsMap), matching
  HealthRepositoryImpl.getExaminationConditions. The previous .asBoolean
  threw on a JsonNull or object/array value and the blanket catch then
  discarded every condition, so a single null entry (plausible in a
  CouchDB-sourced blob) silently blanked the whole conditions line on an
  examination that had conditions. Now each value is evaluated safely and
  true flags survive a malformed sibling.
- isNullOrEmpty -> isNullOrBlank so the blank guard is explicit and skips
  the parse; catch now printStackTrace()'s instead of swallowing silently,
  matching the sibling conditions-blob handler on master.
- Rename MyHealthTest -> HealthExaminationTest and add two cases pinning
  the fix: {"diabetes": true, "asthma": null} and a non-boolean value,
  both asserting diabetes still comes through.

Fixes #16452
