+++ app/src/main/java/org/ole/planet/myplanet/model/HealthExamination.kt
@@ -48,6 +48,30 @@ class HealthExamination {
     }
 
     companion object {
+        /**
+         * Parses an examination's conditions JSON blob and returns the names of the
+         * conditions flagged `true`, joined by `", "`. The conditions field is a JSON
+         * object mapping a condition name to a boolean; a missing or non-boolean value
+         * is treated as `false` (tolerated), matching `HealthRepositoryImpl.getExaminationConditions`.
+         * A null/blank, malformed, or non-object document yields an empty string.
+         */
+        fun formatConditions(conditions: String?): String {
+            if (conditions.isNullOrBlank()) return ""
+            return try {
+                val conditionsMap = JsonUtils.gson.fromJson(conditions, JsonObject::class.java)
+                if (conditionsMap != null) {
+                    conditionsMap.keySet()
+                        .filter { JsonUtils.getBoolean(it, conditionsMap) }
+                        .joinToString(", ")
+                } else {
+                    ""
+                }
+            } catch (e: Exception) {
+                e.printStackTrace()
+                ""
+            }
+        }
+
         fun fromJson(act: JsonObject?): HealthExamination {
             val myHealth = HealthExamination()
             myHealth._id = JsonUtils.getString("_id", act)
