--- app/src/test/java/org/ole/planet/myplanet/model/MyHealthTest.kt
@@ -1,68 +1,0 @@
-package org.ole.planet.myplanet.model
-
-import org.junit.Assert.assertEquals
-import org.junit.Test
-
-class MyHealthTest {
-
-    @Test
-    fun formatConditions_nullReturnsEmpty() {
-        assertEquals("", MyHealth.formatConditions(null))
-    }
-
-    @Test
-    fun formatConditions_emptyReturnsEmpty() {
-        assertEquals("", MyHealth.formatConditions(""))
-    }
-
-    @Test
-    fun formatConditions_blankReturnsEmpty() {
-        assertEquals("", MyHealth.formatConditions("   "))
-    }
-
-    @Test
-    fun formatConditions_onlyTrueFlagsAreJoined() {
-        val conditions = """
-            {
-              "diabetes": true,
-              "hypertension": false,
-              "asthma": true
-            }
-        """.trimIndent()
-
-        val result = MyHealth.formatConditions(conditions).split(", ").toSet()
-
-        assertEquals(setOf("diabetes", "asthma"), result)
-    }
-
-    @Test
-    fun formatConditions_allFalseReturnsEmpty() {
-        val conditions = """{"diabetes": false, "hypertension": false}"""
-
-        assertEquals("", MyHealth.formatConditions(conditions))
-    }
-
-    @Test
-    fun formatConditions_allTrueReturnsAll() {
-        val conditions = """{"diabetes": true, "hypertension": true}"""
-
-        val result = MyHealth.formatConditions(conditions).split(", ").toSet()
-
-        assertEquals(setOf("diabetes", "hypertension"), result)
-    }
-
-    @Test
-    fun formatConditions_malformedJsonReturnsEmpty() {
-        assertEquals("", MyHealth.formatConditions("{not valid json"))
-    }
-
-    @Test
-    fun formatConditions_nonObjectJsonReturnsEmpty() {
-        assertEquals("", MyHealth.formatConditions("[\"diabetes\", \"asthma\"]"))
-    }
-
-    @Test
-    fun formatConditions_emptyObjectReturnsEmpty() {
-        assertEquals("", MyHealth.formatConditions("{}"))
-    }
-}
